requirements.txt
Flask>=3.1
python-dotenv>=1.0
requests>=2.32
SQLAlchemy>=2.0
psycopg[binary]>=3.1  # or: psycopg[binary]>=3.1 and use postgresql+psycopg://
APScheduler>=3.10
tzdata
pytest
flask-login